+++ sdk/resourcemanager/consumption/armconsumption/zz_generated_polymorphic_helpers.go
@@ -49,6 +49,25 @@ func unmarshalChargeSummaryClassificationArray(rawMsg json.RawMessage) ([]Charge
 	return fArray, nil
 }
 
+func unmarshalChargeSummaryClassificationMap(rawMsg json.RawMessage) (map[string]ChargeSummaryClassification, error) {
+	if rawMsg == nil {
+		return nil, nil
+	}
+	var rawMessages map[string]json.RawMessage
+	if err := json.Unmarshal(rawMsg, &rawMessages); err != nil {
+		return nil, err
+	}
+	fMap := make(map[string]ChargeSummaryClassification, len(rawMessages))
+	for key, rawMessage := range rawMessages {
+		f, err := unmarshalChargeSummaryClassification(rawMessage)
+		if err != nil {
+			return nil, err
+		}
+		fMap[key] = f
+	}
+	return fMap, nil
+}
+
 func unmarshalReservationRecommendationClassification(rawMsg json.RawMessage) (ReservationRecommendationClassification, error) {
 	if rawMsg == nil {
 		return nil, nil
@@ -88,6 +107,25 @@ func unmarshalReservationRecommendationClassificationArray(rawMsg json.RawMessag
 	return fArray, nil
 }
 
+func unmarshalReservationRecommendationClassificationMap(rawMsg json.RawMessage) (map[string]ReservationRecommendationClassification, error) {
+	if rawMsg == nil {
+		return nil, nil
+	}
+	var rawMessages map[string]json.RawMessage
+	if err := json.Unmarshal(rawMsg, &rawMessages); err != nil {
+		return nil, err
+	}
+	fMap := make(map[string]ReservationRecommendationClassification, len(rawMessages))
+	for key, rawMessage := range rawMessages {
+		f, err := unmarshalReservationRecommendationClassification(rawMessage)
+		if err != nil {
+			return nil, err
+		}
+		fMap[key] = f
+	}
+	return fMap, nil
+}
+
 func unmarshalUsageDetailClassification(rawMsg json.RawMessage) (UsageDetailClassification, error) {
 	if rawMsg == nil {
 		return nil, nil
@@ -125,4 +163,23 @@ func unmarshalUsageDetailClassificationArray(rawMsg json.RawMessage) ([]UsageDet
 		fArray[index] = f
 	}
 	return fArray, nil
+}
+
+func unmarshalUsageDetailClassificationMap(rawMsg json.RawMessage) (map[string]UsageDetailClassification, error) {
+	if rawMsg == nil {
+		return nil, nil
+	}
+	var rawMessages map[string]json.RawMessage
+	if err := json.Unmarshal(rawMsg, &rawMessages); err != nil {
+		return nil, err
+	}
+	fMap := make(map[string]UsageDetailClassification, len(rawMessages))
+	for key, rawMessage := range rawMessages {
+		f, err := unmarshalUsageDetailClassification(rawMessage)
+		if err != nil {
+			return nil, err
+		}
+		fMap[key] = f
+	}
+	return fMap, nil
 }