+++ sdk/resourcemanager/consumption/armconsumption/zz_generated_usagedetails_client.go
@@ -10,15 +10,15 @@ package armconsumption
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
 	"strconv"
 	"strings"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // UsageDetailsClient contains the methods for the UsageDetails group.
@@ -29,8 +29,15 @@ type UsageDetailsClient struct {
 }
 
 // NewUsageDetailsClient creates a new instance of UsageDetailsClient with the specified values.
-func NewUsageDetailsClient(con *arm.Connection) *UsageDetailsClient {
-	return &UsageDetailsClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewUsageDetailsClient(credential azcore.TokenCredential, options *arm.ClientOptions) *UsageDetailsClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &UsageDetailsClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // List - Lists the usage details for the defined scope. Usage details are available via this API only for May 1, 2014 or later.
@@ -50,9 +57,6 @@ func (client *UsageDetailsClient) List(scope string, options *UsageDetailsListOp
 // listCreateRequest creates the List request.
 func (client *UsageDetailsClient) listCreateRequest(ctx context.Context, scope string, options *UsageDetailsListOptions) (*policy.Request, error) {
 	urlPath := "/{scope}/providers/Microsoft.Consumption/usageDetails"
-	if scope == "" {
-		return nil, errors.New("parameter scope cannot be empty")
-	}
 	urlPath = strings.ReplaceAll(urlPath, "{scope}", scope)
 	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.ep, urlPath))
 	if err != nil {
@@ -84,7 +88,7 @@ func (client *UsageDetailsClient) listCreateRequest(ctx context.Context, scope s
 func (client *UsageDetailsClient) listHandleResponse(resp *http.Response) (UsageDetailsListResponse, error) {
 	result := UsageDetailsListResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.UsageDetailsListResult); err != nil {
-		return UsageDetailsListResponse{}, err
+		return UsageDetailsListResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }